fix: refactored file naming conventions
