+++ src/components/Footer/styles.scss
