+++ src/components/InputError/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import './InputError.scss';
+import './styles.scss';
 
 const InputError = ({ error, errorMessage }: { error: boolean; errorMessage: string }) => (
   <span className='inputError__message'>{error ? errorMessage : ''}</span>
