+++ src/components/Loader/styles.scss
