+++ src/components/Modal/styles.scss
