+++ src/containers/SelectContainer/index.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { INPUT } from 'src/actions/constants';
-import Select from 'src/components/Select/Select';
-import InputError from 'src/components/InputError/InputError';
+import Select from 'src/components/Select';
+import InputError from 'src/components/InputError';
 import { states } from 'src/utils/Helper';
-import { RootState } from 'src/reducers/index';
-import './SelectContainer.scss';
+import { RootState } from 'src/reducers';
+import './styles.scss';
 
 const SelectContainer = ({
   defaultOption,
