+++ src/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { render } from 'react-dom';
 import { Provider } from 'react-redux';
 import store from 'src/store/index';
-import App from 'src/components/App/App';
+import App from 'src/components/App';
 import * as serviceWorker from 'src/serviceWorker';
 import 'src/index.scss';
