+++ src/store/index.tsx
@@ -1,5 +1,5 @@
 import { createStore } from 'redux';
-import rootReducer from 'src/reducers/index';
+import rootReducer from 'src/reducers';
 
 const store = createStore(
   rootReducer,
