+++ src/stories/components/Input.stories.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
-import Input from 'src/components/Input/Input';
+import Input from 'src/components/Input';
 
 const inputProps = {
   error: true,
