+++ src/stories/components/Loader.stories.js
@@ -1,5 +1,5 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
-import Loader from 'src/components/Loader/Loader';
+import Loader from 'src/components/Loader';
 
 storiesOf('Loader', module).add('default', () => <Loader />);
